feat: cmd/ipfs: Make it possible to depend on cmd/ipfs
