+++ cmd/ipfs/kubo/add_migrations.go
@@ -1,4 +1,4 @@
-package main
+package kubo
 
 import (
 	"context"
