+++ cmd/ipfs/kubo/daemon.go
@@ -1,4 +1,4 @@
-package main
+package kubo
 
 import (
 	"errors"
