+++ cmd/ipfs/kubo/daemon_other.go
@@ -1,7 +1,7 @@
 //go:build !linux
 // +build !linux
 
-package main
+package kubo
 
 func notifyReady() {}
