+++ cmd/ipfs/kubo/debug.go
@@ -1,4 +1,4 @@
-package main
+package kubo
 
 import (
 	"net/http"
