+++ cmd/ipfs/kubo/pinmfs_test.go
@@ -1,4 +1,4 @@
-package main
+package kubo
 
 import (
 	"context"
